+++ sdk/identity/azure-identity/test/ut/default_azure_credential_test.cpp
@@ -522,3 +522,9 @@ TEST_P(LogMessagesForSpecificCredential, )
 
   Logger::SetListener(nullptr);
 }
+
+TEST(DefaultAzureCredential, RequireCredentialSpecifierEnvVarValue)
+{
+  EXPECT_THROW(
+      static_cast<void>(std::make_unique<DefaultAzureCredential>(true)), AuthenticationException);
+}